statsd: Fix compiler warning (int32_t vs. int64_t)
Fixes the following compiler warning:

stats_statsd.c: In function ‘osmo_stats_reporter_create_statsd’: stats_statsd.c:54:18: warning: assignment from incompatible pointer type [-Wincompatible-pointer-types]
  srep->send_item = osmo_stats_reporter_statsd_send_item;

+++ include/osmocom/core/stats.h
@@ -78,7 +78,7 @@ struct osmo_stats_reporter {
 	int (*send_item)(struct osmo_stats_reporter *srep,
 		const struct osmo_stat_item_group *statg,
 		const struct osmo_stat_item_desc *desc,
-		int32_t value);
+		int64_t value);
 };
 
 struct osmo_stats_config {
